+++ sdk-extension/opentelemetry-sdk-extension-aws/tests/performance/benchmarks/trace/test_benchmark_aws_xray_ids_generator.py
@@ -12,14 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from opentelemetry.sdk.extension.aws.trace import AwsXRayIdsGenerator
+from opentelemetry.sdk.extension.aws.trace import AwsXRayIdGenerator
 
-ids_generator = AwsXRayIdsGenerator()
+id_generator = AwsXRayIdGenerator()
 
 
 def test_generate_xray_trace_id(benchmark):
-    benchmark(ids_generator.generate_trace_id)
+    benchmark(id_generator.generate_trace_id)
 
 
 def test_generate_xray_span_id(benchmark):
-    benchmark(ids_generator.generate_span_id)
+    benchmark(id_generator.generate_span_id)