+++ sdk-extension/opentelemetry-sdk-extension-aws/tests/trace/test_aws_xray_ids_generator.py
@@ -16,23 +16,23 @@ import datetime
 import time
 import unittest
 
-from opentelemetry.sdk.extension.aws.trace import AwsXRayIdsGenerator
+from opentelemetry.sdk.extension.aws.trace import AwsXRayIdGenerator
 from opentelemetry.trace.span import INVALID_TRACE_ID
 
 
-class AwsXRayIdsGeneratorTest(unittest.TestCase):
+class AwsXRayIdGeneratorTest(unittest.TestCase):
     def test_ids_are_valid(self):
-        ids_generator = AwsXRayIdsGenerator()
+        id_generator = AwsXRayIdGenerator()
         for _ in range(1000):
-            trace_id = ids_generator.generate_trace_id()
+            trace_id = id_generator.generate_trace_id()
             self.assertTrue(trace_id != INVALID_TRACE_ID)
-            span_id = ids_generator.generate_span_id()
+            span_id = id_generator.generate_span_id()
             self.assertTrue(span_id != INVALID_TRACE_ID)
 
     def test_id_timestamps_are_acceptable_for_xray(self):
-        ids_generator = AwsXRayIdsGenerator()
+        id_generator = AwsXRayIdGenerator()
         for _ in range(1000):
-            trace_id = ids_generator.generate_trace_id()
+            trace_id = id_generator.generate_trace_id()
             trace_id_time = trace_id >> 96
             current_time = int(time.time())
             self.assertLessEqual(trace_id_time, current_time)